feat: merge postcss options (#799)
* chore: update version

* fix: mpa entries (#781)

* fix: tabbar state update (#783)

* fix: tabbar state update

* chore: optimize code

* chore: optimize code

* chore: test cases

* chore: optimize code

* chore: optimize code

* chore: optimize component name

* chore: format code

* feat: merge-postcss-options

* feat: merge postcss options
